% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flights2.R
\docType{data}
\name{flights2}
\alias{flights2}
\title{Flights data}
\format{Data frame with columns
\describe{
\item{scheduled}{Scheduled departure time}
\item{dep_time,arr_time}{Departure and arrival times, local tz.}
\item{dep_delay,arr_delay}{Departure and arrival delays, in minutes.
  Negative times represent early departures/arrivals.}
\item{carrier}{Two letter carrier abbreviation. See \code{\link[nycflights13]{airlines}}
  to get name}
\item{tailnum}{Plane tail number}
\item{flight}{Flight number}
\item{origin,dest}{Origin and destination. See \code{\link[nycflights13]{airports}} for
  additional metadata.}
\item{air_time}{Amount of time spent in the air}
\item{distance}{Distance flown}
\item{year,month,day}{Date of departure}
\item{hour,minute}{Time of departure broken in to hour and minutes}
\item{wday}{Day of week of departure}
\item{clock}{Time of departure affixed to the meaningless date 01-01-2000. This is a convenience variable for plotting departures by time of day.}
}}
\source{
RITA, Bureau of transportation statistics,
 \url{http://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}
}
\usage{
flights2
}
\description{
On-time data for all flights that departed NYC (i.e. JFK, LGA or EWR) in
2013.
}
\keyword{datasets}

